requirements.txt
Flask~=3.1.1
Flask-SocketIO~=5.5.1
pyserial~=3.5
uptime
python-can~=4.5.0
esptool~=4.8.1
reedsolo~=1.7.0
requests~=2.32.4
bitstring~=4.3.1
intelhex~=2.3.0
logger~=1.4
rich-click~=1.8.9
PyYAML~=6.0.2
click~=8.2.1
setuptools~=80.9.0
cryptography~=45.0.4
PyAudio~=0.2.14
numpy~=2.3.0
matplotlib~=3.10.3
scipy~=1.15.3
tornado~=6.5.1
websocket-client~=1.8.0
nbformat~=5.10.4